refactor main and functions
