+++ .gitignore
@@ -1,3 +1,4 @@
 *.pyc
 *.pdf
 *.keras
+*.vscode
