--- src/models/keywords-by-term.json
@@ -1,1 +1,0 @@
-{"972": {"1441": 0, "947": 1, "952": 2, "966": 3, "968": 4, "975": 5}, "1441": {"953": 0, "955": 1, "956": 2, "959": 3, "960": 4, "974": 5}, "953": {"949": 0, "950": 1, "957": 2, "961": 3, "963": 4, "970": 5, "971": 6, "976": 7}, "959": {"948": 0, "973": 1}, "974": {"954": 0, "958": 1, "962": 2, "967": 3}, "958": {"949": 0}, "962": {"2315": 0}, "975": {"2316": 0, "41": 1, "44": 2, "964": 3}}
